[REFACTOR] BattleService 위치 변경
